% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttenuatedR.R
\name{getAttenuatedR}
\alias{getAttenuatedR}
\title{Function that gives the attenuated R for an unattenuated R and two reliability values}
\usage{
getAttenuatedR(unattR, rel1, rel2, dp = 3)
}
\arguments{
\item{unattR}{unattenuated R}

\item{rel1}{reliability of first of the variables (order of variables is arbitrary)}

\item{rel2}{reliability of second of the variables (order of variables is arbitrary)}

\item{dp}{number of decimal places required for attenuated R}
}
\value{
numeric: attenuated correlation
}
\description{
This is just the conventional formula for the attenuation of a (Pearson) correlation by unreliability.
}
\section{Background}{

This is ancient psychometrics but still of some use.  For more information, see:
\href{https://www.psyctc.org/psyctc/glossary2/attenuation-by-unreliability-of-measurement/}{OMbook glossary entry for attenuation}
The formula is simple:
\loadmathjax{}

\mjdeqn{attenuatedCorr=unattenuatedCorr*\sqrt{rel_{1}*rel_{2}}}{attenuatedR = unattR * sqrt(rel1 * rel2)}

The short summary is that unreliability in the measurement of both variables involved in a correlation
always reduces the observed correlation between the variables from what it would have been had the
variables been measured with no unreliability (which is essentially impossible for any self-report measures
and pretty much any measures used in our fields.
}

\section{History/development log}{

Started 12.x.24
}

\examples{
getAttenuatedR(.9, .7, .8)



}
\seealso{
Other utility functions:
\code{\link[=convertClipboardAuthorNames]{convertClipboardAuthorNames()}},
\code{\link[=fixVarNames]{fixVarNames()}},
\code{\link[=getCorrectedR]{getCorrectedR()}},
\code{\link[=getLastDateInMonth]{getLastDateInMonth()}},
\code{\link[=getSvalFromPval]{getSvalFromPval()}},
\code{\link[=whichSetOfN]{whichSetOfN()}}
}
\author{
Chris Evans
}
\concept{utility functions}
