% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBootCICorr.R
\name{getBootCICorr}
\alias{getBootCICorr}
\title{Function to return observed correlation between two variables with bootstrap CI}
\usage{
getBootCICorr(
  formula1,
  data,
  method = "p",
  bootReps = 1000,
  conf = 0.95,
  bootCImethod = "pe"
)
}
\arguments{
\item{formula1}{formula defining the two variables to be correlated as var1 ~ var2}

\item{data}{data.frame or tibble with the data, often subset of the data created with group_by() and pick()}

\item{method}{string giving correlation method, can be single letter 'p', 's' or 'k' for pearson, spearman or kendall (in cor())}

\item{bootReps}{integer giving number of bootstrap replications}

\item{conf}{numeric value giving width of confidence interval, e.g. .95 (default)}

\item{bootCImethod}{string giving method to derive bootstrap CI, can be two letters 'pe', 'no', 'ba' or 'bc' for percentile, normal, basic or bca}
}
\value{
list of named values: obsCorr, LCLCorr and UCLCorr
}
\description{
Function to return observed correlation between two variables with bootstrap CI
}
\section{History/development log}{

Started before 5.iv.21
}

\examples{
\dontrun{
library(tidyverse)
### create some data
set.seed(12345) # make this replicable
n <- 150
tibble(x = rnorm(n), y = rnorm(y)) \%>\%
   ### that's got us sample x and y from population in which they're uncorrelated
   ### make them correlated:
   mutate(y = y + .2 * x)  -> data

data \%>\%
   ### don't forget to prefix the call with "list(" to tell dplyr
   ### you are creating list output
   summarise(corr = list(getBootCICorr(x ~ y,
             pick(everything()),
             ### pick(everything()) is, to my mind, a rather verbose replacement for cur_data()
             method = "p", # gets the Pearson correlation
             bootReps = 1000,
             ### "pe" in next line gets the percentile bootstrap CI
             bootCImethod = "pe"))) \%>\%
   ### now unnest the list output to separate columns
   unnest_wider(corr)
}

}
\seealso{
Other bootstrap CI functions:
\code{\link[=getBootCICSC]{getBootCICSC()}},
\code{\link[=getBootCIalpha]{getBootCIalpha()}},
\code{\link[=getBootCIgrpMeanDiff]{getBootCIgrpMeanDiff()}},
\code{\link[=getBootCImean]{getBootCImean()}}
}
\author{
Chris Evans
}
\concept{bootstrap CI functions}
