% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBootCIPearson.R
\name{getBootCIPearson}
\alias{getBootCIPearson}
\title{Title
Function to return observed Pearson correlation between two variables with bootstrap CI}
\usage{
getBootCIPearson(formula1, data, bootReps = 1000, conf = 0.95)
}
\arguments{
\item{formula1}{formula}

\item{data}{data.frame}

\item{bootReps}{integer}

\item{conf}{numeric}
}
\value{
list
}
\description{
Title
Function to return observed Pearson correlation between two variables with bootstrap CI
}
\examples{
\dontrun{
library(tidyverse)
data \%>\%
   summarise(corr = list(getBootCIPearson(var1 ~ var2, cur_data()))) \%>\%
   unnest_wider(corr)
   }
}
