% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCronbachAlpha.R
\name{getChronbachAlpha}
\alias{getChronbachAlpha}
\title{Simple function to return Cronbach's alpha}
\usage{
getChronbachAlpha(dat, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{dat}{data, matrix, data frame or tibble and numeric item data as columns}

\item{na.rm}{logical which causes function to throw error if there is missing data}

\item{verbose}{logical, if FALSE switches off message about n(items) and number of usable rows}
}
\value{
Cronbach as numeric
}
\description{
Simple function to return Cronbach's alpha
}
\section{Background}{

Very simple function that does a fair bit of sanity checking on the input and returns Cronbach's alpha
which was first described comprehensively in the classic paper Cronbach (1951).  There is an extensive
literature criticising alpha as an index of internal consistency/reliability often, to my reading,
failing to note that Cronbach's original paper was quite clear about the issues.  Alpha is not a measure
of unidimensionality but a measure of shared covariance and there are strong arguments that it is not
the best indicator of internal reliability when there are marked differences in variance between items.
The various arguments have led many to recommend McDonald's omega as a better measure than alpha.
Perhaps partly because the canonical reference to McDonald's omega is to his book (McDonald, 1999) and
not to a paper, the literature on this is, to my mind again, a bit complicated.  There is good coverage
in the help for omega() in Revelle's package psych \link[psych]{omega}. Despite the complexities alpha
remains the most reported statistic for internal relability of multi-item measures with some good
reasons: it is simple, robust and never claimed to be doing things it doesn't do!
}

\section{History/development log}{

Started before 5.iv.21
}

\examples{
\dontrun{
### make some nonsense data
tmpMat <- matrix(rnorm(200), ncol = 10)
### default call
getChronbachAlpha(tmpMat)
### switch off message with dimensions of non-missing data
getChronbachAlpha(tmpMat, verbose = FALSE)
}

}
\references{
Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests. Psychometrika, 16(3), 297–334.
McDonald R. P. (1999) Test Theory: A Unified Treatment. Mahwah, NJ: Erlbaum
}
\seealso{
Other Chronbach alpha functions:
\code{\link[=getBootCIalpha]{getBootCIalpha()}}
}
\author{
Chris Evans
}
\concept{Chronbach alpha functions}
