% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrectedR.R
\name{getCorrectedR}
\alias{getCorrectedR}
\title{Function that gives the corrected R for an observed R and two reliability values}
\usage{
getCorrectedR(obsR, rel1, rel2, dp = 3)
}
\arguments{
\item{obsR}{observed R}

\item{rel1}{reliability of first of the variables (order of variables is arbitrary)}

\item{rel2}{reliability of second of the variables (order of variables is arbitrary)}

\item{dp}{number of decimal places required for corrected R}
}
\value{
numeric: corrected correlation
}
\description{
This just uses the conventional formula for the attenuation of a (Pearson) correlation by unreliability.
}
\section{Background}{

This is ancient psychometrics but still of some use.  For more information, see:
\href{https://www.psyctc.org/psyctc/glossary2/attenuation-by-unreliability-of-measurement/}{OMbook glossary entry for attenuation}
The formula is simple:
\loadmathjax{}

\mjdeqn{correctedCorr=\frac{observedCorr}{\sqrt{rel_{1}*rel_{2}}}}{correctedR = obsR / sqrt(rel1 * rel2)}

The short summary is that unreliability in the measurement of both variables involved in a correlation
always reduces the observed correlation between the variables from what it would have been had the
variables been measured with no unreliability (which is essentially impossible for any self-report measures
and pretty much any measures used in our fields.  This uses that relationship to work back to an
assumed "corrected" correlation given an observed correlation and two reliability values.

For even moderately high observed correlations and low reliabilities the function can easily return
values for the corrected correlation over 1.0.  That's a clear indication that things other than
unreliability and classical test theory are at work.  The function gives a warning in this situation.
}

\section{History/development log}{

Started 13.x.24
}

\examples{
getCorrectedR(.3, .7, .7)
### should return 0.428571428571429



}
\seealso{
Other utility functions:
\code{\link[=convertClipboardAuthorNames]{convertClipboardAuthorNames()}},
\code{\link[=fixVarNames]{fixVarNames()}},
\code{\link[=getAttenuatedR]{getAttenuatedR()}},
\code{\link[=getLastDateInMonth]{getLastDateInMonth()}},
\code{\link[=getSvalFromPval]{getSvalFromPval()}},
\code{\link[=whichSetOfN]{whichSetOfN()}}
}
\author{
Chris Evans
}
\concept{utility functions}
