% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNNA.R
\name{getNNA}
\alias{getNNA}
\title{Trivial function to count number of NA values in input}
\usage{
getNNA(x)
}
\arguments{
\item{x}{vector containing the values to check for NAs}
}
\value{
numeric: the number of NA values in
}
\description{
Trivial function to count number of NA values in input
}
\section{History/development log}{

Started before 5.iv.21
}

\examples{
getNNA(c(1:5, NA, 7:10))
getNNA(c(1:5, NA, 7:10))

}
\seealso{
\code{\link{getNOK}} for count of non-missing values

Other counting functions:
\code{\link[=getNOK]{getNOK()}}

Other data checking functions:
\code{\link[=checkAllUnique]{checkAllUnique()}},
\code{\link[=getNOK]{getNOK()}},
\code{\link[=isOneToOne]{isOneToOne()}}
}
\author{
Chris Evans
}
\concept{counting functions}
\concept{data checking functions}
