% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNOK.R
\name{getNOK}
\alias{getNOK}
\title{Trivial function to count number of non-NA values in input}
\usage{
getNOK(x)
}
\arguments{
\item{x}{is a vector of values}
}
\value{
numeric: the number of non-NA values in x
}
\description{
Trivial function to count number of non-NA values in input
}
\section{History/development log}{

Started before 5.iv.21
}

\examples{
getNOK(c(1:5, NA, 7:10))

}
\seealso{
\code{\link{getNNA}} for count of missing values

Other counting functions:
\code{\link[=getNNA]{getNNA()}}

Other data checking functions:
\code{\link[=checkAllUnique]{checkAllUnique()}},
\code{\link[=getNNA]{getNNA()}},
\code{\link[=isOneToOne]{isOneToOne()}}
}
\author{
Chris Evans
}
\concept{counting functions}
\concept{data checking functions}
