% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{isNothing}
\alias{isNothing}
\title{Title Tests for NULL, NA or NAN}
\usage{
isNothing(x)
}
\arguments{
\item{x}{input to test}
}
\value{
logical TRUE if x is NULL, NA or NAN
}
\description{
Title Tests for NULL, NA or NAN
}
\section{Background}{

from https://stackoverflow.com/questions/19655579/a-function-that-returns-true-on-na-null-nan-in-r
function contributed by https://stackoverflow.com/users/2275286/hbat
}

\keyword{internal}
