% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichSetofN.R
\name{whichSetOfN}
\alias{whichSetOfN}
\title{Function that takes an index number and gives which sequential set of N it is in}
\usage{
whichSetOfN(x, n)
}
\arguments{
\item{x}{object to test}

\item{n}{size of set}
}
\value{
integer: the number of the set, 1,2,3 ...
}
\description{
Finds the sequential number of sets of data when reading fixed size multirow blocks of rows.
}
\section{Background}{

I am quite often importing data with a multirow nested structure so I may have data from participants
with different ID values and with different occasions per participant and then some fixed number of
rows of data per person per occasion. For one set of data I might have say four rows of medication
data per participant per occasion per actual medication prescribed.  I use whichSetOfN(row_number(), 4)
to tell me the sequential number of the prescription this row (found by row_number() inside a group_by()).
}

\section{History/development log}{

Started 12.x.24
}

\examples{
whichSetOfN(1:7, 3)
### shows that 1:3 belong to set 1, 4:6 to set 2 and 7 to set 3


}
\seealso{
Other utility functions:
\code{\link[=convertClipboardAuthorNames]{convertClipboardAuthorNames()}},
\code{\link[=fixVarNames]{fixVarNames()}},
\code{\link[=getAttenuatedR]{getAttenuatedR()}},
\code{\link[=getCorrectedR]{getCorrectedR()}},
\code{\link[=getLastDateInMonth]{getLastDateInMonth()}},
\code{\link[=getSvalFromPval]{getSvalFromPval()}}
}
\author{
Chris Evans
}
\concept{utility functions}
