% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CECPfuns-package.R
\docType{package}
\name{CECPfuns-package}
\alias{CECPfuns}
\alias{CECPfuns-package}
\title{CECPfuns: Package of Utility Functions Created by Chris Evans and Clara Paz}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
