% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CECPfuns-package.R
\docType{package}
\name{CECPfuns-package}
\alias{CECPfuns-package}
\alias{CECPfuns}
\title{CECPfuns: Package of Utility Functions for Psychological Therapies, Mental Health and Well-being Work (Created by Chris Evans and Clara Paz)}
\description{
This should evolve into a repository of all the functions that I (CE)
  and Clara Paz (CP) have created (so far only CE!) and tested enough to
  want to store here where we will know we can always find the latest
  versions.  Please check the MIT licence for disclaimer.
}
\details{
Install from github:
\code{remotes::install_github("cpsyctc/CECPfuns")}
or
\code{devtools::install_github("cpsyctc/CECPfuns")}

Run:
\code{vignette("Introduction", package = "CECPfuns")}
to get the the vignette that gives the best introduction to the package
or jump to the index of the package's (as with any R package, index link is at the bottom of every
help page within any package).
}
\author{
\strong{Maintainer}: Chris Evans \email{chris@psyctc.org} (\href{https://orcid.org/0000-0002-4197-4202}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Clara Paz \email{clara.paz@udla.edu.ec} (\href{https://orcid.org/0000-0003-1106-9567}{ORCID}) [consultant, reviewer]
  \item Emily Blackshaw \email{blackshe@roehampton.ac.uk} (\href{https://orcid.org/0000-0002-2674-9189}{ORCID}) [reviewer]
}

}
