% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2CEdate.R
\name{convert2CEdate}
\alias{convert2CEdate}
\title{Title
Function to convert Dates to my date format strings}
\usage{
convert2CEdate(date)
}
\arguments{
\item{date}{Date}
}
\value{
string
}
\description{
Title
Function to convert Dates to my date format strings
}
\examples{
date1 <- as.Date("1/1/2021", format = "\%d/\%m/\%Y")
convert2CEdate(date1)
date2 <- as.Date("2/1/2021", format = "\%d/\%m/\%Y")
dates12 <- c(date1, date2)
convert2CEdate(dates12)
}
