% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVector2sentence.R
\name{convertVector2sentence}
\alias{convertVector2sentence}
\title{Function that converts a vector to an English language list: e.g. 1:3 becomes 1, 2 and 3}
\usage{
convertVector2sentence(x, andVal = " and ", quoted = FALSE, quoteChar = "\\"")
}
\arguments{
\item{x}{input vector to convert}

\item{andVal}{string to put between penultimate and last entry}

\item{quoted}{logical indicating whether each item should be quoted and hence}

\item{quoteChar}{string to put around each item to quote it}
}
\value{
string
}
\description{
Function that converts a vector to an English language list: e.g. 1:3 becomes 1, 2 and 3
}
\examples{
convertVector2sentence(1:4)

}
\seealso{
\code{\link{hyphenateWords}} for another utility function to convert numbers to English words, e.g. "87" to "eighty-seven".

Other text utilities: 
\code{\link{hyphenateWords}()}
}
\author{
Chris Evans
}
\concept{text utilities}
