% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVector2sentence.R
\name{convertVector2sentence}
\alias{convertVector2sentence}
\title{Title
convertVector2sentence}
\usage{
convertVector2sentence(x, andVal = " and ", quoted = FALSE, quoteChar = "\\"")
}
\arguments{
\item{x}{input vector to convert}

\item{andVal}{string to put between penultimate and last entry}

\item{quoted}{logical indicating whether each item should be quoted and hence}

\item{quoteChar}{string to put around each item to quote it}
}
\value{
string
}
\description{
Title
convertVector2sentence
}
\examples{
convertVector2sentence(1:4)
}
