% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSC.R
\name{getCSC}
\alias{getCSC}
\title{Title
Function to return Jacobson (et al.) Clinically Significant Change (CSC)}
\usage{
getCSC(formula1, data)
}
\arguments{
\item{formula1}{defines variables to use as score ~ grp where score has scores and grp has grouping}

\item{data}{is the data, typically cur_data() when called in tidyverse pipe}
}
\value{
single value for the CSC
}
\description{
Title
Function to return Jacobson (et al.) Clinically Significant Change (CSC)
}
\examples{
\dontrun{
# example uses very crude simulated data but illustrates utility of being able to pipe grouped
# data into getCSC()
library(tidyverse)
set.seed(12345)
n <- 250 # total sample size
scores <- rnorm(n) # get some random numbers
# now random allocation to help-seeking or non-help-seeking samples
grp <- sample(c("HS", "notHS"), n, replace = TRUE)
# and random allocation by gender
gender <- sample(c("F", "M"), n, replace = TRUE)
list(scores = scores,
    grp = grp,
    gender = gender) \%>\%
 as_tibble() \%>\%
 mutate(scores = if_else(gender == "F", scores + .13, scores), # make women .13 higher scoring
        ### then make help-seeking group score 1.1 higher
        scores = if_else(grp == "HS", scores + 1.1, scores)) -> tibDat
### now get CSC overall
tibDat \%>\%
 summarise(CSC = getCSC(scores ~ grp, cur_data()))

### get CSC by gender
tibDat \%>\%
 group_by(gender) \%>\%
 summarise(CSC = getCSC(scores ~ grp, cur_data()))
}
}
