% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNNA.R
\name{getNNA}
\alias{getNNA}
\title{Trivial function to count number of NA values in input}
\usage{
getNNA(x)
}
\arguments{
\item{x}{vector containing the values to check for NAs}
}
\value{
numeric: the number of NA values in
}
\description{
Trivial function to count number of NA values in input
}
\examples{
getNNA(c(1:5, NA, 7:10))
getNNA(c(1:5, NA, 7:10))

}
\seealso{
\code{\link{getNOK}} for count of non-missing values

Other counting functions: 
\code{\link{getNOK}()}

Other data checking functions: 
\code{\link{getNOK}()},
\code{\link{isOneToOne}()}
}
\author{
Chris Evans
}
\concept{counting functions}
\concept{data checking functions}
