% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNNA.R
\name{getNNA}
\alias{getNNA}
\title{Title
Trivial function to count number of NA values in input}
\usage{
getNNA(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric
}
\description{
Title
Trivial function to count number of NA values in input
}
\examples{
getNNA(c(1:5, NA, 7:10))
getNNA(c(1:5, NA, 7:10))
}
