% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNOK.R
\name{getNOK}
\alias{getNOK}
\title{Title
Trivial function to count number of non-NA values in input}
\usage{
getNOK(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric
}
\description{
Title
Trivial function to count number of non-NA values in input
}
\examples{
getNOK(c(1:5, NA, 7:10))

}
\seealso{
\code{\link{getNNA}} for count of missing values

Other counting functions: 
\code{\link{getNNA}()}
}
\concept{counting functions}
