% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphenateWords.R
\name{hyphenateWords}
\alias{hyphenateWords}
\title{Title
Function to convert numbers to hyphenated (English) words, e.g. 87 to "eighty-seven"}
\usage{
hyphenateWords(x, capitalise = FALSE)
}
\arguments{
\item{x}{numeric}

\item{capitalise}{logical}
}
\value{
character
}
\description{
Title
Function to convert numbers to hyphenated (English) words, e.g. 87 to "eighty-seven"
}
\examples{
hyphenateWords(87)
hyphenateWords(87, capitalise = TRUE)
}
