% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphenateWords.R
\name{hyphenateWords}
\alias{hyphenateWords}
\title{Function to convert numbers to hyphenated (English) words, e.g. 87 to "eighty-seven"}
\usage{
hyphenateWords(x, capitalise = FALSE)
}
\arguments{
\item{x}{numeric}

\item{capitalise}{logical}
}
\value{
character
}
\description{
Function to convert numbers to hyphenated (English) words, e.g. 87 to "eighty-seven"
}
\section{History/development log}{

Started before 5.iv.21
}

\examples{
hyphenateWords(87)
hyphenateWords(87, capitalise = TRUE)
}
\seealso{
\code{\link{convertVector2sentence}} for another utility function for text, converts vectors to comma (or other symbol), separated clauses, e.g. 1:3 to "1, 2 and 3".

Other text utilities: 
\code{\link{convertVector2sentence}()}
}
\author{
Chris Evans
}
\concept{text utilities}
