% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinconf.R
\name{plotBinconf}
\alias{plotBinconf}
\title{Function outputting a plot of confidence of a proportion for a range of sample sizes}
\usage{
plotBinconf(proportion, conf, minN, maxN, step = 1)
}
\arguments{
\item{proportion}{numeric: the proportion sought (actual proportion will be nearest possible for each n)}

\item{conf}{numeric: confidence interval width, usually .95}

\item{minN}{numeric: the smallest sample size, n, to estimate and plot}

\item{maxN}{numeric: the largest n}

\item{step}{numeric: the steps to use between minN and maxN, defaults to 1 but set higher if plotting a wide range of n}
}
\value{
a ggplot object, by default that will print but you can save it and modify it as you like
}
\description{
Function outputting a plot of confidence of a proportion for a range of sample sizes
}
\examples{
\dontrun{

### 95\% CI around proportion .5 for n from 10 to 70
plotBinconf(.5, .95, 10, 70)

### same exporting to tmpPlot and then changing plot
plotBinconf(.5, .95, 10, 70) -> tmpPlot
tmpPlot +
   ggtitle("95\% CI around proportion .5 for n from 10 to 70") +
   theme_bw()

### other inputs
plotBinconf(0, .95, 10, 70)
plotBinconf(1, .95, 10, 70)
plotBinconf(.7, .95, 100, 200)
plotBinconf(.3, .95, 100, 700, 5)
}

}
