% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantileCIsfromDat.R
\name{plotQuantileCIsfromDat}
\alias{plotQuantileCIsfromDat}
\title{Title Function that plots ecdf of a vector of values with quantiles marked with their CIs}
\usage{
plotQuantileCIsfromDat(
  vecDat = NA,
  vecQuantiles = NA,
  method = c("Nyblom", "Exact", "Bootstrap"),
  ci = 0.95,
  R = 9999,
  type = NULL,
  xLab = "Value for probability (quantile)",
  yLab = "Probability",
  colPoint = "black",
  colLCL = "black",
  colUCL = "black",
  titleText = "ECDF with quantiles and CIs around quantiles",
  titleJust = 0.5,
  themeBW = TRUE,
  printPlot = TRUE,
  returnPlot = FALSE
)
}
\arguments{
\item{vecDat}{Vector of the data}

\item{vecQuantiles}{Vector of quantiles wanted}

\item{method}{Method to compute the CIs}

\item{ci}{Width of CI wanted}

\item{R}{Number of bootstrap replications if bootstrap CI method invoked}

\item{type}{Type of quantile (defaults to 8)}

\item{xLab}{Label for x axis}

\item{yLab}{Label for y axis}

\item{colPoint}{Colour of quantile point (defaults to "black")}

\item{colLCL}{Colour of LCL point (defaults to "black")}

\item{colUCL}{Colour of UCL point (defaults to "black")}

\item{titleText}{Text for title}

\item{titleJust}{Justification for title (0, .5 or 1)}

\item{themeBW}{Logical: whether or not to use them_bw()}

\item{printPlot}{Logical: whether to print the plot (defaults to TRUE)}

\item{returnPlot}{Logical: whether to return the plot (defaults to FALSE)}
}
\value{
The plot as a ggplot object if returnPlot == TRUE
}
\description{
Title Function that plots ecdf of a vector of values with quantiles marked with their CIs
}
\section{Background}{

Quantiles, like percentiles, can be a very useful way of
understanding a distribution, in MH research, typically a
distribution of scores, often either help-seeking sample
distribution or a non-help-seeking sample.

For example, you might want to see if a score you have for
someone is above the 95\% (per)centile, i.e. above the .95
quantile of the non-help-seeking sample distribution.

The catch is that, like any sample statistic, a
quantile/percentile is generally being used as an estimate
of a population value.  That is to say we are really interested
in whether the score we have is above the 95\% percentile of a
non-help-seeking population.  That means that the
quantiles/percentiles from a sample are only approximate guides
to the population values.

As usual, a confidence interval (CI) around an observed quantile
is a way of seeing how precisely, assuming random sampling of
course, the sample quantile can guide us about the population
quantile.

Again as usual, the catch is that there are different ways of
compute the CI around a quantile.  This function is a wrapper
around three methods from Michael Höhle's quantileCI package:
\itemize{
\item the Nyblom method
\item the "exact" method
\item the bootstrap method
}

As I understand the quantileCI package the classic paper by
Nyblom, generally the Nyblom method will be as good or better
than the exact method and much faster than bootstrapping
so the Nyblom method is the default here but I have put the
others in for completeness.
}

\section{History}{

Started 30.vii.23
}

\examples{
\dontrun{
### will need tidyverse to run
library(tidyverse)
### will need quantileCI to run
### if you don't have quantileCI package you need to get it from github:
# devtools::install_github("hoehleatsu/quantileCI")
library(quantileCI)

### use the exact method
getCIforQuantiles(1:1000, c(.1, .5, .95), "e", ci = .95, R = 9999, type = 8)
### use the Nyblom method
getCIforQuantiles(1:1000, c(.1, .5, .95), "n", ci = .95, R = 9999, type = 8)
### use the bootstrap method
getCIforQuantiles(1:1000, c(.1, .5, .95), "b", ci = .95, R = 9999, type = 8)
}

}
\references{
\itemize{
\item Nyblom, J. (1992). Note on interpolated order statistics.
Statistics & Probability Letters, 14(2), 129–131.
https://doi.org/10.1016/0167-7152(92)90076-H
\item https://github.com/mhoehle/quantileCI
}
}
\seealso{
Other CI functions, quantile functions: 
\code{\link{getCIforQuantiles}()}
}
\author{
Chris Evans
}
\concept{CI functions, quantile functions}
